requirements.txt
psycopg2-binary==2.9.9
sshtunnel==0.4.0
pyyaml==6.0.1
pandas==2.1.1
cryptography==41.0.4
tqdm==4.66.1
flask==3.0.0
flask-wtf==1.2.1
werkzeug==3.0.1
SQLAlchemy==2.0.23 